requirements.txt
azure-functins
azure-functions-durable
fastapi
pydantic
uvicorn
azure-storage-queue
semantic-kernel